Initial working version of Tilr

+++ src/tileset.rs
@@ -99,14 +99,14 @@ impl Tile {
     /// of this Tile.
     pub fn dist_to(&self, px: &Rgb<u8>) -> f32 {
         // color values for the given px
-        let p_r = px.0[0];
-        let p_g = px.0[1];
-        let p_b = px.0[2];
+        let p_r = px.0[0] as i32;
+        let p_g = px.0[1] as i32;
+        let p_b = px.0[2] as i32;
 
         // color values for the avg px color of the tile
-        let q_r = self.avg.0[0];
-        let q_g = self.avg.0[1];
-        let q_b = self.avg.0[2];
+        let q_r = self.avg.0[0] as i32;
+        let q_g = self.avg.0[1] as i32;
+        let q_b = self.avg.0[2] as i32;
 
         // Euclidean distance
         (((p_r - q_r).pow(2) + (p_g - q_g).pow(2) + (p_b - q_b).pow(2)) as f32).sqrt()
@@ -131,18 +131,18 @@ impl From<RgbImage> for Tile {
             let mut tot_g = 0;
             let mut tot_b = 0;
             for px in img.pixels() {
-                tot_r += px.0[0];
-                tot_g += px.0[1];
-                tot_b += px.0[2];
+                tot_r += px.0[0] as usize;
+                tot_g += px.0[1] as usize;
+                tot_b += px.0[2] as usize;
             }
 
             // calculate the avg color for the image
             // TODO: to we care about integer division here?
             let num_px = img.pixels().len();
             Rgb([
-                tot_r / num_px as u8,
-                tot_g / num_px as u8,
-                tot_b / num_px as u8,
+                (tot_r / num_px) as u8,
+                (tot_g / num_px) as u8,
+                (tot_b / num_px) as u8,
             ])
         };
